AHA I have found the bug in kodkodparser. see description
OK so parboiled is weird. It parses in two phases: rule construction, and rule execution. During rule construction, each rule function is executed; THE RULE FUNCTIONS ARE NEVER EXECUTED AGAIN. This is some reflection magic. During rule execution, "recursive" expression parsing (i.e. when NaryExpr can parse an NaryExpr argument) isn't actually recursive. the parsing all happens in the context of one function call, the call that was made during rule construction. This means ActionVars are shared between all "calls" to a Rule. So one must be careful that recursive calls do not mutate the state of the current call. This can be accomplished by having action expressions in a rule reset the ActionVars, and restore them just before the rule ends.