add basic indenter
The indenter currently has a few rules (w/ indent witdth 2):
- The first line starts at indent level 0
- Any line ending with a closing brace decreases the indent level,
starting at that line
- Any line ending with an opening brace or bar increases the indent level on the following line
- Otherwise, the indent level of a line matches the indent level of the
previous line

Shortcomings:
- We want to do something smarter on lines following those which end
with a bar. In particular, consider this program

all x : A |
  all y : A - x |
    R[x, y]
some A

The line "some A" should return to the indent level of the highest
binder of the previous line, which in this case is the line "all x : A |".
This doesn't work right now, so either use braces if you want to move
the body of a quantifier to a new line, otherwise keep the body on the
same line as the binder.
- tabify currently isn't called on enter; it should be. So if you write
the following

sig A {
  stuff : A
  }

The indenter *does know* that the closing brace should go back a level,
but it doesn't fire on enter; you have to manually hit tab.