Fix bug with check commands and traces.
Previously, a trace didn't work with a check command, since the check would look for a counterexample to the formula (tracePredicates and checkPredicates), which would find counterexamples where the trace didn't hold. Now, it searches for cases where the trace holds buth the checkPredicates don't (essentially treating the trace as a fact)

Also cleaned up a lot of commented out code/improperly formatted whitespace and added a few comments.