logging: do not gzip check-ex-spec logs
Don't compress files for the check-ex-spec logs. Save the file as a string.

The old design mixes bytes and strings in a terrible way:

1. gzip is a bytes format ... we need to store exactly those bytes to get
   a meaningful result out
2. converting bytes to a string in Racket is a silently lossy transformation
   ... both `string->bytes/utf-8` and `with-output-to-string` generate error
   characters when they find a byte that does not fit UTF-8

We could push bytes to SQL / Postgres, but there may be a similar problem
with their TEXT and MEDIUMTEXT datatypes ... and we surely don't want to
change schemas.