Atom names fixed, my-expand unecessary
In order to make them not unbound id, the expander wraps atom names in (atom .).
To make it easier for the expander to do this, atom names in forge surface need a ` before them.
This menas my-expand is no longer necessary (and it has updated to be the identity function to confirm this).