add the engine_verbosity option
This option is now also documented in the Wiki. It affects how detailed
the logging from the Java engine is. The default is 0 (logging off).

This is a stopgap solution to the problem sometimes encountered by
Windows users, on temporal problems with large numbers of quantifiers,
which was related to Racket-side buffering. The overall problem is not
yet resolved.