Avoid truncating command name sent to Sterling
Previously Sterling would display "orary-name" in the header
where the command is shown. This was due to truncation done
in the engine: removing the first 9 characters of the verbose
command and then removing all blankspace-separated tokens afterward.
E.g. "(run temporary-name ..." would become "orary-name".

Assuming the value of 9 is obsolete, but that just changing it to
5 would possibly cause the same issue later, the code now extracts
the second blankspace-separated token. If there isn't such, it will
just use the full verbose command.